refactor: add stresms

+++ services/files/deleteFile.js
@@ -3,15 +3,18 @@ import path from 'path';
 
 
 export const deleteFile = async (currentDir, commandArguments) => {
-    const fileName = commandArguments[0]
-    const fileToRemovePath = path.join(currentDir, fileName)
-
     try {
+        if (commandArguments.length<2) {
+            throw new Error("You've missed some argument. Please, try once more time.");
+        };
+        const fileName = path.basename(commandArguments[0]);
+        const fileToRemovePath = path.resolve(currentDir, commandArguments[0]);
         const files = await readdir(currentDir);
+
 
         if (files.includes(fileName)) {
             await unlink(fileToRemovePath);
-            console.log(`Operation succeeded.` );
+            console.log(`Operation succeeded.`);
         } else {
             throw new Error(`Operation failed. There's no file ${fileName}`)
         }
